Add license
Add MIT license for code and CC0 for images. MIT allows the code to be
re-used in other use cases. There aren't that many reasons to choose a
more restrictive license such as GPL or AGPL.

CC0 for images was chosen
so that the generated images can be used without referencing
Fyysikkospeksi as the original creator (CC BY would have issues with
maybe having to claim part author in TG bio or somewhere else).